Rename base.py to more descriptive names.
