+++ j5/backends/__init__.py
@@ -1,5 +1,5 @@
 """Backend classes."""
 
-from .base import Backend, CommunicationError, Environment
+from .backend import Backend, CommunicationError, Environment
 
 __all__ = ["Backend", "CommunicationError", "Environment"]
