+++ j5/boards/__init__.py
@@ -1,5 +1,5 @@
 """This module contains the boards that we support."""
 
-from .base import Board, BoardGroup
+from .board import Board, BoardGroup
 
 __all__ = ["Board", "BoardGroup"]
