+++ j5/components/__init__.py
@@ -1,6 +1,6 @@
 """This module contains components, which are the smallest logical element of hardware."""
 
-from .base import Component, Interface, NotSupportedByHardwareError
+from .component import Component, Interface, NotSupportedByHardwareError
 from .battery_sensor import BatterySensor, BatterySensorInterface
 from .button import Button, ButtonInterface
 from .gpio_pin import GPIOPin, GPIOPinInterface, GPIOPinMode
